backend/pyproject.toml
[tool.poetry]
name = "livewater-backend"
version = "0.1.0"
description = "Livewater backend service"
authors = ["Alex Bausk <bauskas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.10.3"
google-cloud-bigquery = "^1.16"
requests = "^2.22"
python-dateutil = "^2.8"
gunicorn = "^19.9"
aiohttp = "^3.5"
cchardet = "^2.1"
aiodns = "^2.0"
aiohttp-session = "^2.7"
aiohttp-security = "^0.4.0"
cryptography = "^2.7"
aiohttp_cors = "^0.7.0"

[tool.poetry.dev-dependencies]
ptvsd = "^4.2"
pylint = "^2.3"
rope = "^0.14.0"
autopep8 = "^1.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
